Move Wrapper and Button classes to sdl_wrappper. Change platform_specific.fixup_window functionality to return pygame surface for Screen
